+++ src/Ext/Sidebar/Body.cpp
@@ -5,6 +5,9 @@
 #include <HouseClass.h>
 #include <SuperClass.h>
 
+#include <Ext/Rules/Body.h>
+#include <Ext/BuildingType/Body.h>
+
 std::unique_ptr<SidebarExt::ExtData> SidebarExt::Data = nullptr;
 
 SHPStruct* SidebarExt::TabProducingProgress[4];
@@ -52,12 +55,20 @@ bool __stdcall SidebarExt::AresTabCameo_RemoveCameo(BuildType* pItem)
 	if (pItem->ItemType == AbstractType::BuildingType || pItem->ItemType == AbstractType::Building)
 	{
 		__assume(pTechnoType != nullptr);
-		buildCat = static_cast<BuildingTypeClass*>(pTechnoType)->BuildCat;
-		auto& pDisplay = DisplayClass::Instance;
-		pDisplay.SetActiveFoundation(nullptr);
-		pDisplay.CurrentBuilding = nullptr;
-		pDisplay.CurrentBuildingType = nullptr;
-		pDisplay.CurrentBuildingOwnerArrayIndex = -1;
+		// It is not necessary to remove buildings on the mouse in all cases here
+		const auto pBldType = static_cast<BuildingTypeClass*>(pTechnoType);
+		buildCat = pBldType->BuildCat;
+		const auto pDisplay = &DisplayClass::Instance;
+		const auto pCurType = abstract_cast<BuildingTypeClass*>(pDisplay->CurrentBuildingType);
+
+		if (!RulesExt::Global()->ExtendedBuildingPlacing || !pCurType
+			|| BuildingTypeExt::IsSameBuildingType(pBldType, pCurType))
+		{
+			pDisplay->SetActiveFoundation(nullptr);
+			pDisplay->CurrentBuilding = nullptr;
+			pDisplay->CurrentBuildingType = nullptr;
+			pDisplay->CurrentBuildingOwnerArrayIndex = -1;
+		}
 	}
 
 	// AbandonAll contains Abandon, if the factory cannot be found, it will also cannot be found when respont to this event.