+++ src/Ext/Techno/Body.h
@@ -54,6 +54,7 @@ public:
 		int WHAnimRemainingCreationInterval;
 		WeaponTypeClass* LastWeaponType;
 		CellClass* FiringObstacleCell; // Set on firing if there is an obstacle cell between target and techno, used for updating WaveClass target etc.
+		CDTimerClass UnitAutoDeployTimer;
 		bool IsDetachingForCloak; // Used for checking animation detaching, set to true before calling Detach_All() on techno when this anim is attached to and to false after when cloaking only.
 		int BeControlledThreatFrame;
 		DWORD LastTargetID;
@@ -120,6 +121,7 @@ public:
 			, WHAnimRemainingCreationInterval { 0 }
 			, LastWeaponType {}
 			, FiringObstacleCell {}
+			, UnitAutoDeployTimer {}
 			, IsDetachingForCloak { false }
 			, BeControlledThreatFrame { 0 }
 			, LastTargetID { 0xFFFFFFFF }