+++ src/Ext/TerrainType/Hooks.Passable.cpp
@@ -6,11 +6,6 @@
 
 #include <Utilities/GeneralUtils.h>
 
-constexpr bool IS_CELL_OCCUPIED(CellClass* pCell)
-{
-	return pCell->OccupationFlags & 0x20 || pCell->OccupationFlags & 0x40 || pCell->OccupationFlags & 0x80 || pCell->GetInfantry(false);
-}
-
 // Passable TerrainTypes Hook #1 - Do not set occupy bits.
 DEFINE_HOOK(0x71C110, TerrainClass_SetOccupyBit_PassableTerrain, 0x6)
 {
@@ -85,63 +80,4 @@ DEFINE_HOOK(0x73FB71, UnitClass_CanEnterCell_PassableTerrain, 0x6)
 	}
 
 	return 0;
-}
-
-// Buildable-upon TerrainTypes Hook #1 - Allow placing buildings on top of them.
-// DEFINE_HOOK(0x73FEC1, UnitClass_WhatAction_DeploysIntoDesyncFix, 0x6) in Hooks.DeploysInto.cpp
-
-// Buildable-upon TerrainTypes Hook #2 - Draw laser fence placement even if they are on the way.
-DEFINE_HOOK(0x6D57C1, TacticalClass_DrawLaserFencePlacement_BuildableTerrain, 0x9)
-{
-	enum { ContinueChecks = 0x6D57D2, DontDraw = 0x6D59A6 };
-
-	GET(CellClass*, pCell, ESI);
-
-	if (auto const pTerrain = pCell->GetTerrain(false))
-		return TerrainTypeExt::ExtMap.Find(pTerrain->Type)->CanBeBuiltOn ? ContinueChecks : DontDraw;
-
-	return ContinueChecks;
-}
-
-// Buildable-upon TerrainTypes Hook #3 - Remove them when buildings are placed on them.
-DEFINE_HOOK(0x5684B1, MapClass_PlaceDown_BuildableTerrain, 0x6)
-{
-	GET(ObjectClass*, pObject, EDI);
-	GET(CellClass*, pCell, EAX);
-
-	if (pObject->WhatAmI() == AbstractType::Building)
-	{
-		if (auto const pTerrain = pCell->GetTerrain(false))
-		{
-			if (TerrainTypeExt::ExtMap.Find(pTerrain->Type)->CanBeBuiltOn)
-			{
-				pCell->RemoveContent(pTerrain, false);
-				TerrainTypeExt::Remove(pTerrain);
-			}
-		}
-	}
-
-	return 0;
-}
-
-// Buildable-upon TerrainTypes Hook #4 -> Allow placing walls on top of terrain
-DEFINE_HOOK(0x5FD2B6, OverlayClass_Unlimbo_SkipTerrainCheck, 0x9)
-{
-	enum { Unlimbo = 0x5FD2CA, NoUnlimbo = 0x5FD2C3 };
-
-	GET(CellClass* const, pCell, EAX);
-
-	if (!Game::IsActive)
-		return Unlimbo;
-
-	if (auto const pTerrain = pCell->GetTerrain(false))
-	{
-		if (!TerrainTypeExt::ExtMap.Find(pTerrain->Type)->CanBeBuiltOn)
-			return NoUnlimbo;
-
-		pCell->RemoveContent(pTerrain, false);
-		TerrainTypeExt::Remove(pTerrain);
-	}
-
-	return Unlimbo;
 }